# apps/api-go/api/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/courses:
    get:
      operationId: CoursesController_findAll_v1
      parameters:
      - name: sort
        required: false
        in: query
        schema:
          enum:
          - newest
          - recent
          - rating
          - popular
          type: string
      - name: search
        required: false
        in: query
        description: 标题/描述/讲师模糊搜索
        schema: {}
      - name: courseType
        required: false
        in: query
        description: 课程类型过滤
        schema:
          enum:
          - own
          - partner
          - public
          - third_party
          type: string
      - name: status
        required: false
        in: query
        description: 课程状态过滤
        schema:
          enum:
          - draft
          - published
          - archived
          type: string
      responses:
        '200':
          description: ''
      summary: 获取课程列表（公开）
      tags:
      - courses
    post:
      operationId: CoursesController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCourseDto'
      responses:
        '201':
          description: ''
      security:
      - bearer: []
      summary: 创建课程（管理员）
      tags:
      - courses
  /api/v1/courses/{id}:
    get:
      operationId: CoursesController_findOne_v1
      parameters:
      - name: id
        required: true
        in: path
        description: 课程ID
        schema:
          type: string
      responses:
        '200':
          description: ''
      summary: 获取课程详情（公开；管理员可见草稿）
      tags:
      - courses
    patch:
      operationId: CoursesController_update_v1
      parameters:
      - name: id
        required: true
        in: path
        description: 课程ID
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCourseDto'
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 更新课程（管理员）
      tags:
      - courses
    delete:
      operationId: CoursesController_delete_v1
      parameters:
      - name: id
        required: true
        in: path
        description: 课程ID
        schema:
          type: string
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 删除课程（管理员）
      tags:
      - courses
...
  /api/v1/courses/{courseId}/chapters:
    get:
      operationId: ChaptersController_list_v1
      parameters:
      - name: courseId
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 列出课程章节(含 lessons + resources)
      tags:
      - chapters
    post:
      operationId: ChaptersController_create_v1
      parameters:
      - name: courseId
        required: true
        in: path
        schema:
          type: string
      responses:
        '201':
          description: ''
      security:
      - bearer: []
      summary: 新建章节
      tags:
      - chapters
  /api/v1/courses/chapters/{id}:
    patch:
      operationId: ChaptersController_update_v1
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 改章节
      tags:
      - chapters
    delete:
      operationId: ChaptersController_remove_v1
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 软删章节(同时级联软删 lessons)
      tags:
      - chapters
  /api/v1/courses/{courseId}/chapters/reorder:
    post:
      operationId: ChaptersController_reorder_v1
      parameters:
      - name: courseId
        required: true
        in: path
        schema:
          type: string
      responses:
        '201':
          description: ''
      security:
      - bearer: []
      summary: 批量重排章节顺序
      tags:
      - chapters
...
  /api/v1/badges:
    get:
      operationId: BadgesController_findAll_v1
      parameters: []
      responses:
        '200':
          description: ''
      summary: 获取全部启用徽章（公开）
      tags:
      - badges
    post:
      operationId: BadgesController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBadgeDto'
      responses:
        '201':
          description: ''
      security:
      - bearer: []
      summary: 创建徽章（管理员）
      tags:
      - badges
  /api/v1/badges/me:
    get:
      operationId: BadgesController_getMyBadges_v1
      parameters: []
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 获取我的徽章墙（含进度）
      tags:
      - badges
  /api/v1/badges/{id}:
    patch:
      operationId: BadgesController_update_v1
      parameters:
      - name: id
        required: true
        in: path
        description: 徽章ID
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBadgeDto'
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 更新徽章（管理员）
      tags:
      - badges
    delete:
      operationId: BadgesController_delete_v1
      parameters:
      - name: id
        required: true
        in: path
        description: 徽章ID
        schema:
          type: string
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 删除徽章（管理员）
      tags:
      - badges
  /api/v1/badges/admin/stats:
    get:
      operationId: BadgesController_getAdminStats_v1
      parameters: []
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 管理员数据看板
      tags:
      - badges
...
  /api/v1/admin/stats:
    get:
      operationId: AdminController_getStats_v1
      parameters: []
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 后台看板统计(KPI / 图表 / 待办 / 系统状态)
      tags:
      - admin
...
info:
  title: AI Academy API
  description: AI and LLM education platform API
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateResourceDto:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
          description: 资源 URL (http/https)
        type:
          type: string
          enum:
          - pdf
          - code
          - link
          - video
          - audio
        isLocked:
          type: boolean
          default: true
      required:
      - title
      - url
      - type
    NestedCourseLessonDto:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        videoUrl:
          type: string
          description: 课时视频 URL (http/https)
        videoDuration:
          type: number
        isPreview:
          type: boolean
          default: false
        orderIndex:
          type: number
          description: 排序索引
        resources:
          description: 关联资源
          type: array
          items:
            $ref: '#/components/schemas/CreateResourceDto'
      required:
      - title
    CreateChapterDto:
      type: object
      properties:
        title:
          type: string
          description: 章节标题
        description:
          type: string
          description: 章节描述
        orderIndex:
          type: number
          description: 排序索引
        lessons:
          description: 章节下的课时
          type: array
          items:
            $ref: '#/components/schemas/NestedCourseLessonDto'
      required:
      - title
      - orderIndex
    CreateCourseDto:
      type: object
      properties:
        title:
          type: string
          description: 课程标题
        description:
          type: string
          description: 课程描述
        learningPoints:
          type: string
          description: 学习要点
        instructor:
          type: string
          description: 讲师名称
        instructorId:
          type: string
          description: 讲师用户 UUID
        level:
          type: string
          enum:
          - Beginner
          - Intermediate
          - Advanced
          - Expert
          description: 难度等级
        duration:
          type: string
          description: 学习时长（人类可读，例如 "8h"）
        thumbnail:
          type: string
          description: 课程缩略图 URL
        tags:
          type: string
          description: 标签（逗号分隔或 JSON 字符串）
        costType:
          type: string
          enum:
          - free
          - paid
          - charity
          description: 计费类型
        price:
          type: number
          description: 价格（元，0 表示免费）
        status:
          type: string
          enum:
          - draft
          - published
          - archived
          description: 课程状态
        courseType:
          type: string
          enum:
          - own
          - partner
          - public
          - third_party
          description: 课程类型
        externalUrl:
          type: string
          description: 外链课程 URL（外部跳转类型课程用）
          maxLength: 500
        sourceVideoUrl:
          type: string
          description: 原始视频源 URL（导入用）
          maxLength: 500
        sourcePlatform:
          type: string
          description: 来源平台标识
          maxLength: 20
        industryId:
          type: string
          description: 行业 UUID
        categoryId:
          type: string
          description: 课程分类 UUID
        chapters:
          description: 课程章节（含课时和资源）
          type: array
          items:
            $ref: '#/components/schemas/CreateChapterDto'
      required:
      - title
      - description
      - learningPoints
      - instructor
      - level
      - duration
      - thumbnail
      - tags
      - costType
      - price
    UpdateCourseDto:
      type: object
      properties:
        title:
          type: string
          description: 课程标题
        description:
          type: string
          description: 课程描述
        learningPoints:
          type: string
          description: 学习要点
        instructor:
          type: string
          description: 讲师名称
        instructorId:
          type: string
          description: 讲师用户 UUID
        level:
          type: string
          enum:
          - Beginner
          - Intermediate
          - Advanced
          - Expert
          description: 难度等级
        duration:
          type: string
          description: 学习时长（人类可读，例如 "8h"）
        thumbnail:
          type: string
          description: 课程缩略图 URL
        tags:
          type: string
          description: 标签（逗号分隔或 JSON 字符串）
        costType:
          type: string
          enum:
          - free
          - paid
          - charity
          description: 计费类型
        price:
          type: number
          description: 价格（元，0 表示免费）
        status:
          type: string
          enum:
          - draft
          - published
          - archived
          description: 课程状态
        courseType:
          type: string
          enum:
          - own
          - partner
          - public
          - third_party
          description: 课程类型
        externalUrl:
          type: string
          description: 外链课程 URL（外部跳转类型课程用）
          maxLength: 500
        sourceVideoUrl:
          type: string
          description: 原始视频源 URL（导入用）
          maxLength: 500
        sourcePlatform:
          type: string
          description: 来源平台标识
          maxLength: 20
        industryId:
          type: string
          description: 行业 UUID
        categoryId:
          type: string
          description: 课程分类 UUID
        chapters:
          description: 课程章节（含课时和资源）
          type: array
          items:
            $ref: '#/components/schemas/CreateChapterDto'
      required:
      - title
      - description
      - learningPoints
      - instructor
      - level
      - duration
      - thumbnail
      - tags
      - costType
      - price
...
    CreateBadgeDto:
      type: object
      properties:
        code:
          type: string
          description: 程序内唯一标识，如 first_course
        name:
          type: string
        description:
          type: string
        icon:
          type: string
          description: Lucide 图标名
        category:
          type: string
          description: 分组：learning / streak / milestone ...
        criteriaType:
          type: string
          enum:
          - course_completed
          - lessons_completed
          - streak_days
          - first_enrollment
          - practice_completed
          - points_reached
          - course_specific
        criteriaValue:
          type: number
        points:
          type: number
          description: 解锁时奖励积分
        isActive:
          type: boolean
        orderIndex:
          type: number
      required:
      - code
      - name
      - description
      - criteriaType
    UpdateBadgeDto:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        category:
          type: string
        criteriaType:
          type: string
          enum:
          - course_completed
          - lessons_completed
          - streak_days
          - first_enrollment
          - practice_completed
          - points_reached
          - course_specific
        criteriaValue:
          type: number
        points:
          type: number
        isActive:
          type: boolean
        orderIndex:
          type: number
...